pyproject.toml
[project]
name = "agentic-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.12"
dependencies = [
    "click>=8.3.1",
    "faster-whisper>=1.2.1",
    "llm>=0.28",
    "llm-openrouter>=0.5",
    "pyyaml>=6.0",
]

[project.scripts]
recording-processor = "recording_recording_processor:cli"
toc-generator = "automations.toc_generator.toc_generator:main"
